Restructure modules related to symbolic expressions manipulation
